pyproject.toml
[project]
name = "quark-assistant"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"setuptools>=72.2.0",
	"python-dotenv>=1.0.1",
	"uvicorn>=0.30.6",
	"fastapi>=0.112.1",
	# "haystack-ai" 2.13 does not support Agents in a backward-compatible way
	"haystack-ai>=2.4.0,<2.13",
	"pyyaml>=6.0.2",
	"amazon-bedrock-haystack>=1.0.1",
	"haystack-experimental>=0.1.1",
	"opensearch-haystack>=0.9.0",
	"xmltodict>=0.13.0",
	"pdfminer-six>=20240706",
	"sentence-transformers>=3.0.1",
	"biopython>=1.84",
	"eutils>=0.6.0",
	"python-multipart>=0.0.9",
	"dataclass-wizard>=0.22.3",
	"datargs>=1.1.0",
	"opensearch-py>=2.7.1",
	"google-vertex-haystack>=2.0.0",
	"s3fs>=2024.9.0",
	"arxiv>=2.1.3",
	"amazon-bedrock-haystack>=2.1.0",
	"pysocks>=1.7.1",
	"botocore>=1.35.36",
	"boto3>=1.35.36",
	"pandas>=2.2.3",
	"spacy>=3.8.5",
]

[tool.uv]
index-strategy = "unsafe-best-match"
extra-index-url = ["https://download.pytorch.org/whl/cpu"]

override-dependencies = [
	"torch==2.5.1 ; sys_platform == 'darwin'",  # https://github.com/astral-sh/uv/issues/5945
	"torch==2.5.1+cpu ; sys_platform !='darwin'",  # https://github.com/astral-sh/uv/issues/5945
	"markupsafe<3",  # markupsafe>=3 cannot be installed with torch+cpu versions
]
dev-dependencies = [
    "awscurl>=0.36",
    "ruff>=0.11.6",
]

[tool.black]
line-length = 88
exclude = '''
^/(
  (
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
    | .notebooks
    | .nox
  )
)
'''

[tool.isort]
profile = "black"
extend_skip = [".tmp", "tmp", ".venv"]

[tool.mypy]
exclude = "^.tmp/"
ignore_missing_imports = true

[tool.bandit]
skips = ['B101']

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
	"E",    # pycodestyle
	"F",    # Pyflakes
	"UP",   # pyupgrade
	"B",    # flake8-bugbear
	"SIM",  # flake8-simplify
	"I",    # isort
	"D",    # pydocstyle
	"C401", # flake8-comprehensions: unnecessary-generator-set
	"C402", # flake8-comprehensions: unnecessary-generator-dict
	"C403", # flake8-comprehensions: unnecessary-list-comprehension-set
	"C404", # flake8-comprehensions: unnecessary-list-comprehension-dict
	"C405", # flake8-comprehensions: unnecessary-literal-set
	"W605", # pycodestyle: invalid-escape-sequence
]

# Ignore D100 Missing docstring in public module
ignore = ["D100"]
[tool.ruff.lint.per-file-ignores]
# Also ignore `D104` in all `__init__.py` files.
"__init__.py" = ["D104"]
"tests/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pytest.ini_options]
minversion = "7"
testpaths = ["tests"]
log_cli_level = "INFO"
xfail_strict = true
addopts = ["-ra", "--strict-config", "--strict-markers"]
filterwarnings = ["error", "ignore::DeprecationWarning"]
